--- apps/meteor/app/lib/client/index.ts
@@ -1,2 +1,0 @@
-import '../lib/MessageTypes';
-import './OAuthProxy';
