+++ apps/meteor/client/importPackages.ts
@@ -1,2 +1,1 @@
-import '../app/lib/client';
 import '../app/utils/client';
