+++ apps/meteor/client/meteor/overrides/index.ts
@@ -1,11 +1,12 @@
 import './absoluteUrl';
 import './ddpOverREST';
 import './ddpSdkCollectionBridge';
-import './subscribeViaSDK';
-import './stubMeteorStream';
-import './killMeteorStream';
 import './desktopInjection';
+import './killMeteorStream';
+import './oauthProxy';
 import './oauthRedirectUri';
 import './settings';
+import './stubMeteorStream';
+import './subscribeViaSDK';
 import './totpOnCall';
 import './userAndUsers';
